+++ src/Pages/Public/ConsumerDefense/ConsumerDefensePage.js
@@ -1,0 +1,12 @@
+import WebsiteLayout from "../../../Layouts/WebsiteLayout"
+import ConsumerDefenseBlock from "./ConsumerDefenseBlock"
+
+function ConsumerDefensePage() {
+  return (
+    <WebsiteLayout>
+      <ConsumerDefenseBlock/>
+    </WebsiteLayout>
+  )
+}
+
+export default ConsumerDefensePage
