--- src/Pages/Public/DefensaConsumidor/DefensaConsumidorPage.js
@@ -1,13 +1,0 @@
-import React from "react"
-import WebsiteLayout from "../../../Layouts/WebsiteLayout"
-import DefensaConsumidorBlock from "./DefensaConsumidorBlock"
-
-function DefensaConsumidorPage() {
-  return (
-    <WebsiteLayout>
-      <DefensaConsumidorBlock/>
-    </WebsiteLayout>
-  )
-}
-
-export default DefensaConsumidorPage
